day 2

+++ labs/week02-lab/example/test-3.py
@@ -3,8 +3,8 @@ print("   - Ask for weight (kg) and height (m)")
 print("   - Calculate: BMI = weight / (height ** 2)")
 print()
 
-w = int(input("weight: "))
-h = int(input("height: "))
+w = float(input("weight: "))
+h = float(input("height: "))
 
 bmi = w / (h ** 2)
